rename a few functions.
FrozenSurface is broken. We need to decided the features to project
to before compiling. Thus we need to pick a set of features
before freezing.

This is to say if we'd like to save some stored planck GMM models
we need to freeze a lot.

Unless we work out the way to marginalize a GMM model.

Given that the GMM model in scikit-learn is more or less a black box
we probably have to write our own in order to do this.