Refactor and add Freeze.
I am against marginalizing GMM. It is simple, but estimating
confidence requires the original MCMC chain -- unless we want to
go to rasterize the GMM to a mesh (which is ugly in its own regard.)

So as a solution I am freezing all 1d and 2d projections. This may
be too slow. We'll see.